Elliptical hearts
To demonstrate the gestalt principle of continuity I plot two ellipses
of dots at 90 degrees to each other that I describe using elliptical
equations in an x-y plane. I colour some of the dots in each ellipse
blue and some orange so that the arrangement creates the illusion that
there are two hearts, one the right way round and one upside-down.